Merge Recording methods into MockServer
It wasn't previously possible to return or store the `MockServer` and
`Recording` together because `Recording` borrowed `MockServer` with a
lifetime and prevented it from being moved out of the function:

    error[E0515]: cannot return value referencing local variable `server`
       --> tests/examples/record_and_playback_tests.rs:235:13
        |
    227 |             let recording = server.record(|rule| {
        |                             ------ `server` is borrowed here
    ...
    235 |             Self { server, recording }
        |             ^^^^^^^^^^^^^^^^^^^^^^^^^^ returns a value referencing data owned by the current function

    error[E0505]: cannot move out of `server` because it is borrowed
       --> tests/examples/record_and_playback_tests.rs:235:20
        |
    218 |         fn new(target_url: String) -> Self {
        |                                       ---- return type is RecordingServer<'1>
    219 |             let server = MockServer::start();
        |                 ------ binding `server` declared here
    ...
    227 |             let recording = server.record(|rule| {
        |                             ------ borrow of `server` occurs here
    ...
    235 |             Self { server, recording }
        |             -------^^^^^^-------------
        |             |      |
        |             |      move out of `server` occurs here
        |             returning this value requires that `server` is borrowed for `'1`

Make this possible by moving the methods on `Recording` up to
`MockServer` and just carrying a `RecordingID` around. I've prefixed the
method names with `record_` to indicate that they're not operating on
mocks and `recording_` seemed to long but I'm open to other suggestions.

This also demonstrates writing the recordings when the struct goes
out-of-scope but we can't easily (without renaming from tempfiles)
exercise playback from that path without a deterministic name per issue
115.