Add delete_after_count method to Mock
In situations where a component is expected to automatically retry an
endpoint manually calling `delete` on a `Mock` may not be an option.

To support this scenario, add a new `delete_after_count` method to `When`
that will remove the `Mock` from the server after it has matched the
specified number of calls.

I set the method to panic if a count of 0 is set, as potentially
deleting the mock before triggering it significantly complicated state
management, and logically the caller should not have created the mock in
the first place. I considered having the method take a `NonZeroUsize`,
but I didn't think the added ceremony was worth it.